requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
pymongo==4.13.0
python-dotenv==1.1.0
chardet==5.2.0